Re-organize some components
