+++ mathesar_ui/src/components/AppHeader.svelte
