+++ mathesar_ui/src/components/breadcrumb/BreadcrumbItem.svelte
