+++ mathesar_ui/src/components/breadcrumb/BreadcrumbSeparatorIcon.svelte
