+++ mathesar_ui/src/components/breadcrumb/SchemaSelector.svelte
